Add gunicorn. Fix werkzeug issue.
The server now runs with gunicorn, which is more "prod ready" than the
default web server provided by flask. We can now tune the number of
worker to see if it has impact on perf.
Also fix agressivly the werkzeug version because the version 1 or
werkzeug is not compatible with flask-restplus. In order to upgrade
restplus, we need to switch to flask-restx because restplus in
unmaintained. So prefer a quick fix.